Optionally show examples menu by using SHOW_EXAMPLES environment.
